+++ src/SMAPI/Events/OneSecondUpdateTickingEventArgs.cs
@@ -1,5 +1,5 @@
 using System;
-using StardewValley;
+using StardewModdingAPI.Framework;
 
 namespace StardewModdingAPI.Events
 {
@@ -9,8 +9,8 @@ namespace StardewModdingAPI.Events
         /*********
         ** Accessors
         *********/
-        /// <summary>The number of ticks elapsed since the game started, including the current tick.</summary>
-        public uint Ticks => (uint)Game1.ticks;
+        /// <summary>The number of ticks elapsed since the game started, excluding the upcoming tick.</summary>
+        public uint Ticks => SGame.TicksElapsed;
 
 
         /*********